refactor(test): Split monolithic E2E tests into granular test classes
Refactor the End-to-End test suite by splitting large, monolithic test files (`EndToEndM1.kt` and `EndToEndM2.kt`) into smaller, focused test classes. This improves test maintainability, readability, and allows for better isolation of test scenarios.

**Changes:**
*   Deleted `EndToEndM1.kt` and `EndToEndM2.kt`.
*   Created specific test classes for Milestone 1 scenarios:
    *   `M1Test1CompleteUserFlowTest.kt`: Covers the complete user profile flow (login, edit, logout).
    *   `M1Test2ProfileEditUserFlowTest.kt`: Covers profile edit cancellation.
    *   `M1Test3ProfileEditUserFlowTest.kt`: Covers profile data persistence across navigation.
*   Created specific test classes for Milestone 2 scenarios:
    *   `M2Test1EventUserFlowTest.kt`: Covers event management flow.
    *   `M2Test2FriendUserFowTest.kt`: Covers friend request and management flow.
    *   `M2Test3ChatUserFlowTest.kt`: Covers chat and conversation flow.
    *   `M2Test4MemoryUserFlowTest.kt`: Covers memory creation and management flow.
    *   `M2Test5MapSearchUserFlowTest.kt`: Covers map search and navigation flow.
*   Preserved the global `ReentrantLock` mechanism across the new classes to prevent static mock conflicts during parallel execution.
