--- app/src/androidTest/java/com/swent/mapin/e2e/EndToEndM2.kt
@@ -1,718 +1,0 @@
-package com.swent.mapin.e2e
-
-import android.content.Context
-import androidx.activity.ComponentActivity
-import androidx.compose.ui.test.*
-import androidx.compose.ui.test.junit4.createAndroidComposeRule
-import androidx.navigation.compose.rememberNavController
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.google.android.gms.tasks.Tasks
-import com.google.firebase.FirebaseApp
-import com.google.firebase.Timestamp
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.firestore.CollectionReference
-import com.google.firebase.firestore.DocumentReference
-import com.google.firebase.firestore.DocumentSnapshot
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.QuerySnapshot
-import com.swent.mapin.model.FriendRequest
-import com.swent.mapin.model.FriendshipStatus
-import com.swent.mapin.model.UserProfile
-import com.swent.mapin.model.event.Event
-import com.swent.mapin.model.location.Location
-import com.swent.mapin.model.memory.Memory
-import com.swent.mapin.navigation.AppNavHost
-import com.swent.mapin.navigation.Route
-import com.swent.mapin.testing.UiTestTags
-import io.mockk.*
-import java.util.concurrent.locks.ReentrantLock
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import org.junit.After
-import org.junit.Before
-import org.junit.FixMethodOrder
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-import org.junit.runners.MethodSorters
-
-/**
- * End-to-End Tests for Milestone 2 (M2)
- *
- * This comprehensive test suite includes all E2E tests from both M1 and M2:
- *
- * FROM M1:
- * - User Profile Management (login, edit, save, logout)
- * - Profile edit cancellation
- * - Profile data persistence
- *
- * NEW IN M2:
- * - Event Management (create, view, filter events)
- * - Friends Management (search, send requests, accept/reject)
- * - Chat Flow (navigate, create conversations, send messages)
- * - Memory Management (create memories, link to events, manage visibility)
- * - Map & Search (location search, directions, map layers, filters)
- *
- * All tests use MockK to mock Firebase services and run sequentially via a global lock to prevent
- * conflicts with static mocks.
- */
-@OptIn(ExperimentalCoroutinesApi::class)
-@RunWith(AndroidJUnit4::class)
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-class EndToEndM2 {
-
-  @get:Rule val composeTestRule = createAndroidComposeRule<ComponentActivity>()
-  private lateinit var navController: androidx.navigation.NavHostController
-
-  private lateinit var context: Context
-  private lateinit var mockAuth: FirebaseAuth
-  private lateinit var mockUser: FirebaseUser
-  private lateinit var mockFirestore: FirebaseFirestore
-
-  // Test user data
-  private val testId = System.currentTimeMillis().toString() + "-" + Thread.currentThread().id
-  private val testUserId = "test-user-e2e-$testId"
-  private val testEmail = "testuser@example.com"
-  private val testUserName = "Test User"
-
-  // Profile test data (M1)
-  private val initialBio = "I love testing and coding!"
-  private val updatedName = "Updated Test User"
-  private val updatedBio = "This is my updated bio after editing"
-  private var testProfile: UserProfile = UserProfile()
-
-  // Event test data (M2)
-  private val testEventId = "test-event-$testId"
-  private val testEventTitle = "Test Concert"
-  private val testEventDescription = "An amazing music concert in the park"
-  private val testEventLocation = Location.from("Central Park", 40.7829, -73.9654)
-  private val testEventTags = listOf("Music", "Outdoor")
-  private var testEvent: Event = Event()
-
-  // Friends test data (M2)
-  private val friendUserId = "friend-user-$testId"
-  private val friendUserName = "Friend User"
-  private val requestId = "request-$testId"
-  private var friendUserProfile: UserProfile = UserProfile()
-  private var testFriendRequest: FriendRequest = FriendRequest()
-
-  // Chat test data (M2)
-  private val conversationId = "conversation-$testId"
-  private val testMessage = "Hello, this is a test message!"
-
-  // Memory test data (M2)
-  private val testMemoryId = "test-memory-$testId"
-  private val testMemoryTitle = "Amazing Day at the Festival"
-  private val testMemoryDescription = "Had an incredible time at the festival with friends"
-  private var testMemory: Memory = Memory()
-
-  // Map search test data (M2)
-  private val searchQuery = "Central Park"
-  private val searchLocation = Location.from("Central Park, New York", 40.7829, -73.9654)
-  private val renderMapInTests = false // Disable Mapbox rendering on CI emulators
-
-  companion object {
-    private val globalLock = ReentrantLock()
-    @Volatile private var isLocked = false
-  }
-
-  @Before
-  fun setup() {
-    globalLock.lock()
-    isLocked = true
-
-    try {
-      context = ApplicationProvider.getApplicationContext()
-
-      if (FirebaseApp.getApps(context).isEmpty()) {
-        FirebaseApp.initializeApp(context)
-      }
-
-      // Setup test data
-      testProfile =
-          UserProfile(
-              userId = testUserId,
-              name = testUserName,
-              bio = initialBio,
-              hobbies = listOf("Testing", "Coding"),
-              location = "Test City",
-              avatarUrl = null,
-              bannerUrl = null,
-              hobbiesVisible = true)
-
-      testEvent =
-          Event(
-              uid = testEventId,
-              title = testEventTitle,
-              description = testEventDescription,
-              location = testEventLocation,
-              tags = testEventTags,
-              date = Timestamp.now(),
-              endDate = Timestamp.now(),
-              public = true,
-              ownerId = testUserId,
-              participantIds = emptyList(),
-              capacity = 100)
-
-      friendUserProfile =
-          UserProfile(userId = friendUserId, name = friendUserName, bio = "Friend user bio")
-
-      testFriendRequest =
-          FriendRequest(
-              requestId = requestId,
-              fromUserId = testUserId,
-              toUserId = friendUserId,
-              timestamp = Timestamp.now(),
-              status = FriendshipStatus.PENDING)
-
-      testMemory =
-          Memory(
-              uid = testMemoryId,
-              title = testMemoryTitle,
-              description = testMemoryDescription,
-              eventId = testEventId,
-              ownerId = testUserId,
-              isPublic = false,
-              createdAt = Timestamp.now(),
-              mediaUrls = emptyList(),
-              taggedUserIds = emptyList())
-
-      clearAllMocks()
-
-      // Mock FirebaseAuth
-      mockkStatic(FirebaseAuth::class)
-      mockAuth = mockk(relaxed = true)
-      mockUser = mockk(relaxed = true)
-
-      every { FirebaseAuth.getInstance() } returns mockAuth
-      every { mockAuth.currentUser } returns mockUser
-      every { mockUser.uid } returns testUserId
-      every { mockUser.displayName } returns testUserName
-      every { mockUser.email } returns testEmail
-      every { mockUser.photoUrl } returns null
-      every { mockAuth.signOut() } just Runs
-
-      // Mock Firestore
-      mockkStatic(FirebaseFirestore::class)
-      mockFirestore = mockk(relaxed = true)
-      every { FirebaseFirestore.getInstance() } returns mockFirestore
-
-      // Setup mock collections and documents
-      setupFirestoreMocks()
-    } catch (e: Exception) {
-      if (isLocked && globalLock.isHeldByCurrentThread) {
-        globalLock.unlock()
-        isLocked = false
-      }
-      throw e
-    }
-  }
-
-  private fun setupFirestoreMocks() {
-    // Users collection
-    val mockUsersCollection = mockk<CollectionReference>(relaxed = true)
-    val mockUserDocument = mockk<DocumentReference>(relaxed = true)
-    val mockUserSnapshot = mockk<DocumentSnapshot>(relaxed = true)
-
-    every { mockFirestore.collection("users") } returns mockUsersCollection
-    every { mockUsersCollection.document(any()) } returns mockUserDocument
-    every { mockUserSnapshot.exists() } returns true
-    every { mockUserSnapshot.toObject(UserProfile::class.java) } answers { testProfile }
-    every { mockUserDocument.get() } answers { Tasks.forResult(mockUserSnapshot) }
-    every { mockUserDocument.set(any()) } answers
-        {
-          val profile = firstArg<UserProfile>()
-          testProfile = profile
-          Tasks.forResult(null)
-        }
-    every { mockUserDocument.update(any<String>(), any()) } answers { Tasks.forResult(null) }
-
-    // Events collection
-    val mockEventsCollection = mockk<CollectionReference>(relaxed = true)
-    val mockEventDocument = mockk<DocumentReference>(relaxed = true)
-    val mockEventSnapshot = mockk<DocumentSnapshot>(relaxed = true)
-    val mockQuerySnapshot = mockk<QuerySnapshot>(relaxed = true)
-
-    every { mockFirestore.collection("events") } returns mockEventsCollection
-    every { mockEventsCollection.document(any()) } returns mockEventDocument
-    every { mockEventsCollection.document() } returns mockEventDocument
-    every { mockEventDocument.id } returns testEventId
-    every { mockEventSnapshot.exists() } returns true
-    every { mockEventSnapshot.toObject(Event::class.java) } answers { testEvent }
-    every { mockEventDocument.get() } answers { Tasks.forResult(mockEventSnapshot) }
-    every { mockEventDocument.set(any()) } answers { Tasks.forResult(null) }
-    every { mockEventDocument.update(any<String>(), any()) } answers { Tasks.forResult(null) }
-    every { mockQuerySnapshot.documents } returns listOf(mockEventSnapshot)
-    every { mockEventsCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-
-    // Friend requests collection
-    val mockFriendRequestsCollection = mockk<CollectionReference>(relaxed = true)
-    val mockRequestDocument = mockk<DocumentReference>(relaxed = true)
-    val mockRequestSnapshot = mockk<DocumentSnapshot>(relaxed = true)
-
-    every { mockFirestore.collection("friendRequests") } returns mockFriendRequestsCollection
-    every { mockFriendRequestsCollection.document(any()) } returns mockRequestDocument
-    every { mockFriendRequestsCollection.document() } returns mockRequestDocument
-    every { mockRequestDocument.id } returns requestId
-    every { mockRequestSnapshot.exists() } returns true
-    every { mockRequestSnapshot.toObject(FriendRequest::class.java) } answers { testFriendRequest }
-    every { mockRequestDocument.get() } answers { Tasks.forResult(mockRequestSnapshot) }
-    every { mockRequestDocument.set(any()) } answers { Tasks.forResult(null) }
-    every { mockRequestDocument.update(any<Map<String, Any>>()) } answers { Tasks.forResult(null) }
-    every { mockRequestDocument.delete() } answers { Tasks.forResult(null) }
-    every { mockFriendRequestsCollection.whereEqualTo(any<String>(), any()) } returns
-        mockFriendRequestsCollection
-    every { mockFriendRequestsCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-
-    // Conversations and messages collections
-    val mockConversationsCollection = mockk<CollectionReference>(relaxed = true)
-    val mockMessagesCollection = mockk<CollectionReference>(relaxed = true)
-    val mockConversationDocument = mockk<DocumentReference>(relaxed = true)
-    val mockMessageDocument = mockk<DocumentReference>(relaxed = true)
-    val mockConversationSnapshot = mockk<DocumentSnapshot>(relaxed = true)
-
-    every { mockFirestore.collection("conversations") } returns mockConversationsCollection
-    every { mockFirestore.collection("messages") } returns mockMessagesCollection
-    every { mockConversationsCollection.document(any()) } returns mockConversationDocument
-    every { mockConversationsCollection.document() } returns mockConversationDocument
-    every { mockMessagesCollection.document(any()) } returns mockMessageDocument
-    every { mockMessagesCollection.document() } returns mockMessageDocument
-    every { mockConversationDocument.id } returns conversationId
-    every { mockConversationDocument.collection("messages") } returns mockMessagesCollection
-    every { mockConversationSnapshot.exists() } returns true
-    every { mockConversationSnapshot.id } returns conversationId
-    every { mockConversationDocument.get() } answers { Tasks.forResult(mockConversationSnapshot) }
-    every { mockConversationDocument.set(any()) } answers { Tasks.forResult(null) }
-    every { mockMessageDocument.set(any()) } answers { Tasks.forResult(null) }
-    every { mockConversationsCollection.whereArrayContains(any<String>(), any()) } returns
-        mockConversationsCollection
-    every { mockConversationsCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-    every { mockMessagesCollection.orderBy(any<String>()) } returns mockMessagesCollection
-    every { mockMessagesCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-
-    // Memories collection
-    val mockMemoriesCollection = mockk<CollectionReference>(relaxed = true)
-    val mockMemoryDocument = mockk<DocumentReference>(relaxed = true)
-    val mockMemorySnapshot = mockk<DocumentSnapshot>(relaxed = true)
-
-    every { mockFirestore.collection("memories") } returns mockMemoriesCollection
-    every { mockMemoriesCollection.document(any()) } returns mockMemoryDocument
-    every { mockMemoriesCollection.document() } returns mockMemoryDocument
-    every { mockMemoryDocument.id } returns testMemoryId
-    every { mockMemorySnapshot.exists() } returns true
-    every { mockMemorySnapshot.toObject(Memory::class.java) } answers { testMemory }
-    every { mockMemoryDocument.get() } answers { Tasks.forResult(mockMemorySnapshot) }
-    every { mockMemoryDocument.set(any()) } answers { Tasks.forResult(null) }
-    every { mockMemoryDocument.update(any<Map<String, Any>>()) } answers { Tasks.forResult(null) }
-    every { mockMemoriesCollection.whereEqualTo(any<String>(), any()) } returns
-        mockMemoriesCollection
-    every { mockMemoriesCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-
-    every { mockUsersCollection.whereEqualTo(any<String>(), any()) } returns mockUsersCollection
-    every { mockUsersCollection.get() } answers { Tasks.forResult(mockQuerySnapshot) }
-  }
-
-  @After
-  fun tearDown() {
-    try {
-      unmockkAll()
-    } finally {
-      if (isLocked && globalLock.isHeldByCurrentThread) {
-        globalLock.unlock()
-        isLocked = false
-      }
-    }
-  }
-
-  // ==================== M1 TESTS ====================
-
-  /**
-   * M1 Test: Complete User Profile Management Flow Tests login, edit profile, save, navigate to
-   * settings, and logout.
-   */
-  @Test
-  fun m1_completeUserFlow_loginEditProfileAndLogout_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    // Verify on map screen
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Navigate to profile
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-    composeTestRule.runOnIdle { navController.navigate(Route.Profile.route) }
-    composeTestRule.waitForIdle()
-
-    // Verify on profile screen
-    composeTestRule.onNodeWithTag("profileScreen", useUnmergedTree = true).assertExists()
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      composeTestRule
-          .onAllNodesWithText(testUserName, useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-
-    // Enter edit mode
-    composeTestRule.onNodeWithTag("editButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    // Edit profile
-    composeTestRule.onNodeWithTag("editNameField", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("editNameField", useUnmergedTree = true).performTextClearance()
-    composeTestRule
-        .onNodeWithTag("editNameField", useUnmergedTree = true)
-        .performTextInput(updatedName)
-
-    composeTestRule.onNodeWithTag("editBioField", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("editBioField", useUnmergedTree = true).performTextClearance()
-    composeTestRule
-        .onNodeWithTag("editBioField", useUnmergedTree = true)
-        .performTextInput(updatedBio)
-
-    // Save changes
-    composeTestRule.onNodeWithTag("saveButton", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("saveButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    // Verify changes persist
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      val nodes =
-          composeTestRule
-              .onAllNodesWithText(updatedName, useUnmergedTree = true)
-              .fetchSemanticsNodes()
-      nodes.isNotEmpty()
-    }
-    composeTestRule.onNodeWithText(updatedName, useUnmergedTree = true).assertExists()
-    composeTestRule.onNodeWithText(updatedBio, useUnmergedTree = true).assertExists()
-
-    // Navigate to Settings
-    composeTestRule.runOnIdle { navController.navigate(Route.Settings.route) }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.waitUntil(timeoutMillis = 5000) {
-      composeTestRule
-          .onAllNodesWithTag("settingsScreen", useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-
-    // Logout - scroll to button as Settings screen is scrollable
-    composeTestRule.onNodeWithTag("logoutButton", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("logoutButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      composeTestRule
-          .onAllNodesWithTag("logoutConfirmButton", useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-    composeTestRule.onNodeWithTag("logoutConfirmButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    verify { mockAuth.signOut() }
-
-    // Verify redirect to login
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      composeTestRule
-          .onAllNodesWithTag(UiTestTags.AUTH_SCREEN, useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-    composeTestRule
-        .onNodeWithTag(UiTestTags.AUTH_SCREEN, useUnmergedTree = true)
-        .assertIsDisplayed()
-
-    println("M1: User profile E2E test completed successfully")
-  }
-
-  /** M1 Test: Profile Edit Cancellation Verifies that canceling profile edits discards changes. */
-  @Test
-  fun m1_profileEdit_cancelButton_shouldDiscardChanges() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    // Navigate to profile
-    composeTestRule.runOnIdle { navController.navigate(Route.Profile.route) }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      composeTestRule
-          .onAllNodesWithText(testUserName, useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-
-    // Enter edit mode
-    composeTestRule.onNodeWithTag("editButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    // Edit fields
-    composeTestRule.onNodeWithTag("editNameField", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("editNameField", useUnmergedTree = true).performTextClearance()
-    composeTestRule
-        .onNodeWithTag("editNameField", useUnmergedTree = true)
-        .performTextInput("Temporary Name")
-
-    // Cancel
-    composeTestRule.onNodeWithTag("cancelButton", useUnmergedTree = true).performScrollTo()
-    composeTestRule.onNodeWithTag("cancelButton", useUnmergedTree = true).performClick()
-    composeTestRule.waitForIdle()
-
-    // Verify original data restored
-    composeTestRule.waitUntil(timeoutMillis = 10000) {
-      composeTestRule
-          .onAllNodesWithText(testUserName, useUnmergedTree = true)
-          .fetchSemanticsNodes()
-          .isNotEmpty()
-    }
-    composeTestRule.onNodeWithText("Temporary Name", useUnmergedTree = true).assertDoesNotExist()
-
-    println("M1: Profile edit cancellation test completed successfully")
-  }
-
-  // ==================== M2 TESTS ====================
-
-  /** M2 Test: Event Management Flow Tests creating and interacting with events. */
-  @Test
-  fun m2_completeEventFlow_createViewAndInteract_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Try to create event (button may not exist in current UI)
-    val addEventButtonTags = listOf("addEventButton", "createEventButton", "fabAddEvent")
-    for (tag in addEventButtonTags) {
-      try {
-        val nodes =
-            composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-        if (nodes.isNotEmpty()) {
-          composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-          composeTestRule.waitForIdle()
-          break
-        }
-      } catch (e: Exception) {
-        // Continue
-      }
-    }
-
-    composeTestRule.waitForIdle()
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    println("M2: Event management E2E test completed successfully")
-  }
-
-  /** M2 Test: Friends Management Flow Tests searching for users and managing friend requests. */
-  @Test
-  fun m2_completeFriendsFlow_searchSendRequestAndManage_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Navigate to Friends
-    val friendsButtonTags = listOf("friendsButton", "navigateToFriendsButton")
-    var navigated = false
-    for (tag in friendsButtonTags) {
-      try {
-        val nodes =
-            composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-        if (nodes.isNotEmpty()) {
-          composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-          navigated = true
-          break
-        }
-      } catch (e: Exception) {
-        // Continue
-      }
-    }
-
-    if (!navigated) {
-      composeTestRule.runOnIdle { navController.navigate(Route.Friends.route) }
-    }
-
-    composeTestRule.waitForIdle()
-    Thread.sleep(1000)
-    composeTestRule.waitForIdle()
-
-    // Navigate back to map (optional - may already be on map)
-    try {
-      val backButtonTags = listOf("navigateBackButton", "backButton", "back")
-      for (tag in backButtonTags) {
-        try {
-          val nodes =
-              composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-          if (nodes.isNotEmpty()) {
-            composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-            break
-          }
-        } catch (e: Exception) {
-          // Continue
-        }
-      }
-    } catch (e: Exception) {
-      // Navigation may not be needed
-    }
-
-    composeTestRule.waitForIdle()
-
-    println("M2: Friends management E2E test completed successfully")
-  }
-
-  /** M2 Test: Chat Flow Tests navigating to chat and conversation management. */
-  @Test
-  fun m2_completeChatFlow_navigateCreateAndMessage_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Navigate to Chat
-    val chatButtonTags = listOf("chatButton", "navigateToChatButton")
-    var navigated = false
-    for (tag in chatButtonTags) {
-      try {
-        val nodes =
-            composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-        if (nodes.isNotEmpty()) {
-          composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-          navigated = true
-          break
-        }
-      } catch (e: Exception) {
-        // Continue
-      }
-    }
-
-    if (!navigated) {
-      composeTestRule.runOnIdle { navController.navigate(Route.Chat.route) }
-    }
-
-    composeTestRule.waitForIdle()
-    Thread.sleep(1000)
-    composeTestRule.waitForIdle()
-
-    // Navigate back to map (optional - may already be on map)
-    try {
-      val backButtonTags = listOf("navigateBackButton", "backButton", "back")
-      for (tag in backButtonTags) {
-        try {
-          val nodes =
-              composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-          if (nodes.isNotEmpty()) {
-            composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-            break
-          }
-        } catch (e: Exception) {
-          // Continue
-        }
-      }
-    } catch (e: Exception) {
-      // Navigation may not be needed
-    }
-
-    composeTestRule.waitForIdle()
-
-    println("M2: Chat flow E2E test completed successfully")
-  }
-
-  /** M2 Test: Memory Management Flow Tests creating and managing memories. */
-  @Test
-  fun m2_completeMemoryFlow_createAndManageMemory_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Memory creation would happen through event details
-    // Just verify map is accessible
-    composeTestRule.waitForIdle()
-
-    // Verify map is still accessible
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    println("M2: Memory management E2E test completed successfully")
-  }
-
-  /** M2 Test: Map and Search Flow Tests location search and map interactions. */
-  @Test
-  fun m2_completeMapSearchFlow_searchNavigateAndFilter_shouldWorkCorrectly() {
-    composeTestRule.setContent {
-      navController = rememberNavController()
-      AppNavHost(
-          navController = navController,
-          isLoggedIn = true,
-          renderMap = renderMapInTests,
-          autoRequestPermissions = false)
-    }
-    composeTestRule.waitForIdle()
-
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    // Try to interact with search
-    val searchBarTags = listOf("searchBar", "locationSearchBar", "mapSearchBar")
-    for (tag in searchBarTags) {
-      try {
-        val nodes =
-            composeTestRule.onAllNodesWithTag(tag, useUnmergedTree = true).fetchSemanticsNodes()
-        if (nodes.isNotEmpty()) {
-          composeTestRule.onNodeWithTag(tag, useUnmergedTree = true).performClick()
-          composeTestRule.waitForIdle()
-          break
-        }
-      } catch (e: Exception) {
-        // Continue
-      }
-    }
-
-    composeTestRule.waitForIdle()
-    composeTestRule.onNodeWithTag(UiTestTags.MAP_SCREEN, useUnmergedTree = true).assertExists()
-
-    println("M2: Map and search E2E test completed successfully")
-  }
-}
